Added some minor security fixes

+++ backend/controllers/user.controller.js
@@ -48,10 +48,15 @@ export const updateUser = async (req, res) => {
 
 export const deleteUser = async (req, res) => {
     const { id } = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "User not found" });
+    }
+
     try {
         await User.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "User deleted successfully" });
     } catch (error) {
-        res.status(404).json({ success: false, message: "User not found" });
+        res.status(500).json({ success: false, message: "Internal Error" });
     }
 }
